requirements.txt
botbuilder-core
botbuilder-dialogs
flask